Rename components for consistency
