+++ src/Components/SmartComponents/CveDetail/CveImagesTable.js
@@ -9,7 +9,7 @@ import {
 import { useDispatch, useSelector } from 'react-redux';
 import NoMatchingItems from '../../PresentationalComponents/EmptyStates/NoMatchingItems';
 import { useParams } from 'react-router-dom';
-import BaseTable from '../Generic/BaseTable';
+import BaseTable from '../Table/BaseTable';
 import { useExport, useUrlBoundParams } from '../../../Helpers/hooks';
 import {
   changeCveImagesTableColumns,
